File structure fix - move files related to ffprobe reports to utils directory
